nixos: fix issue where loopbackIps does not work with ip tables
With the following settings:

```
  networking.nat = {
    enable = true;
    internalInterfaces = [ "lo" "eno2" ];
    externalInterface = "eno2";
    forwardPorts = [
      {
        sourcePort = 80;
        proto = "tcp";
        destination = "192.168.100.11:80";
        loopbackIps = [ "127.0.0.1" ];
      }
    ];
  };

  networking.nftables.enable = true;
```

Then `curl 127.0.0.1` connects to the service running on `192.168.100.11:80`.

However, when nftables is not enabled and iptables is in use `curl 127.0.0.1` times out.

This is because nftables enables masquerading. Enabling it here allows the loopbackIps to work as intended.

Likely this is also the cause of user confusion in these examples

https://discourse.nixos.org/t/nat-and-port-forwarding/28929
https://github.com/NixOS/nixpkgs/issues/28721
https://discourse.nixos.org/t/struggling-to-forward-a-localhost-port-to-another-port-on-localhost/28278

and why [this wiki page](https://nixos.wiki/wiki/Networking) has an example that requires enabling nftables and manually forwarding po>